feature: market hub auction & shop

+++ src/main/java/com/accsaber/backend/model/dto/request/item/CreateItemRequest.java
@@ -30,7 +30,7 @@ public class CreateItemRequest {
     private boolean welcomeGrant = false;
     private boolean missionPoolable = false;
     private boolean active = true;
-    private BigDecimal worth;
+    private Long worth;
     private String requirement;
     private Integer unlockLevel;
 }
