+++ src/main/java/com/accsaber/backend/model/dto/request/item/UpdateItemRequest.java
@@ -20,7 +20,7 @@ public class UpdateItemRequest {
     private Boolean stackable;
     private Boolean welcomeGrant;
     private Boolean missionPoolable;
-    private BigDecimal worth;
+    private Long worth;
     private String requirement;
     private Integer unlockLevel;
 }
